add jake's affordability data to our model
As a proxy for a city's affordability, we use median home price of
the predominant surrounding county.

Data are pulled from the following Census endpoint through a
one-time synchronous call to:

   https://api.census.gov/data/2017/acs/acs5/profile?get=DP04_0089E,NAME&for=county:*

The data are post-processed from an array of arrays into a simple object, using county and state
as the key for median home price.

The unit tests provided at the end of assets/js/affordability.js sketch out how
this object can be used to cook down various data sources to just the data needed to drive
our ranking algorithm.