fix spelling errors and factual info on researchers in readme tweak whitespace in reporting political results to be more mobile-friendly
Did more proof reading and caught a missing word and misspelled word.
Also, the researchers merely add commentary to the report and didn't
publish the 182-city ranking.  So fix that detail.

I'm seeing an ugly character-wrap on mobile with the republican '%' sign.
It will still look fine to have the '%' for both dem and rep data directly
adjacent to the numbers.  This should prevent the wrap i'm seeing on iphone SE.